+++ plugins/set.ts
@@ -18,6 +18,7 @@ function setTag(
   }
 
   const expression = code.replace(/^set\s+/, "");
+  const { dataVarname } = env.options;
 
   // Value is set (e.g. {{ set foo = "bar" }})
   if (expression.includes("=")) {
@@ -30,12 +31,12 @@ function setTag(
     const [, variable, value] = match;
     const val = env.compileFilters(tokens, value);
 
-    return `if (__data.hasOwnProperty("${variable}")) {
+    return `if (${dataVarname}.hasOwnProperty("${variable}")) {
       ${variable} = ${val};
     } else {
       var ${variable} = ${val};
     }
-    __data["${variable}"] = ${variable};
+    ${dataVarname}["${variable}"] = ${variable};
     `;
   }
 
@@ -43,7 +44,7 @@ function setTag(
   const compiled: string[] = [];
   const compiledFilters = env.compileFilters(tokens, expression);
 
-  compiled.push(`if (__data.hasOwnProperty("${expression}")) {
+  compiled.push(`if (${dataVarname}.hasOwnProperty("${expression}")) {
     ${expression} = "";
   } else {
     var ${expression} = "";
@@ -58,6 +59,6 @@ function setTag(
 
   tokens.shift();
   compiled.push(`${expression} = ${compiledFilters};`);
-  compiled.push(`__data["${expression.trim()}"] = ${expression};`);
+  compiled.push(`${dataVarname}["${expression.trim()}"] = ${expression};`);
   return compiled.join("\n");
 }